Move GW term management code to the GuardianWeeklySubscriptionBuilder class
