+++ support-workers/src/main/scala/com/gu/zuora/subscriptionBuilders/GuardianWeeklySubscriptionBuilder.scala
@@ -5,13 +5,13 @@ import java.util.UUID
 import com.gu.i18n.Country
 import com.gu.support.config.TouchPointEnvironment
 import com.gu.support.promotions.{DefaultPromotions, PromoCode, PromoError, PromotionService}
-import com.gu.support.workers.{BillingPeriod, GuardianWeekly, SixWeekly}
-import com.gu.support.workers.exceptions.BadRequestException
-import com.gu.support.zuora.api.{ReaderType, SubscriptionData}
-import com.gu.zuora.subscriptionBuilders.ProductSubscriptionBuilders.{applyPromoCode, buildProductSubscription, validateRatePlan}
-import org.joda.time.{DateTimeZone, LocalDate}
-import com.gu.support.workers.ProductTypeRatePlans._
 import com.gu.support.workers.GuardianWeeklyExtensions._
+import com.gu.support.workers.ProductTypeRatePlans._
+import com.gu.support.workers.exceptions.BadRequestException
+import com.gu.support.workers.{BillingPeriod, GuardianWeekly, SixWeekly}
+import com.gu.support.zuora.api.{Day, Month, ReaderType, SubscriptionData}
+import com.gu.zuora.subscriptionBuilders.ProductSubscriptionBuilders.{applyPromoCode, buildProductSubscription, validateRatePlan}
+import org.joda.time.{DateTimeZone, Days, LocalDate}
 
 import scala.util.{Failure, Success, Try}
 
@@ -39,13 +39,20 @@ object GuardianWeeklySubscriptionBuilder {
       guardianWeekly.introductoryRatePlan(environment).map(_.id).getOrElse(recurringProductRatePlanId)
     } else recurringProductRatePlanId
 
+    val (initialTerm, autoRenew, initialTermPeriodType) = if (readerType == ReaderType.Gift)
+      (initialTermInDays(contractEffectiveDate, contractAcceptanceDate, guardianWeekly.billingPeriod.monthsInPeriod), false, Day)
+    else
+      (12, true, Month)
+
     val subscriptionData = buildProductSubscription(
       requestId,
       recurringProductRatePlanId,
       contractAcceptanceDate = contractAcceptanceDate,
       contractEffectiveDate = contractEffectiveDate,
       readerType = readerType,
-      initialTermMonths = guardianWeekly.billingPeriod.monthsInPeriod
+      autoRenew = autoRenew,
+      initialTerm = initialTerm,
+      initialTermPeriodType = initialTermPeriodType
     )
 
     applyPromoCode(promotionService, maybePromoCode, country, promotionProductRatePlanId, subscriptionData)
@@ -53,4 +60,9 @@ object GuardianWeeklySubscriptionBuilder {
 
   private[this] def isIntroductoryPromotion(billingPeriod: BillingPeriod, maybePromoCode: Option[PromoCode]) =
     maybePromoCode.contains(DefaultPromotions.GuardianWeekly.NonGift.sixForSix) && billingPeriod == SixWeekly
+
+  def initialTermInDays(contractEffectiveDate: LocalDate, contractAcceptanceDate: LocalDate, termLengthMonths: Int): Int = {
+    val termEnd = contractAcceptanceDate.plusMonths(termLengthMonths)
+    Days.daysBetween(contractEffectiveDate, termEnd).getDays
+  }
 }
